Start rewriting the parser
The start of the parser rewrite, there are few issues with this new version that will need to be overcome but the code is far more clean than the previous version. Hover is still in an awkard state as the sections need to be decoupled from the chatson parser helper most likely. There are also a few issues with decorators that can be found by running the unit tests.